[WIP] New tokenizer.
